refactor prompts
