+++ listen-interface/src/prompts/index.ts
@@ -1,0 +1,2 @@
+export { systemPromptEvm } from "./evm";
+export { systemPromptSolana } from "./solana";
